+++ packages/spotlight/src/container.js
@@ -11,7 +11,7 @@ import {coerceArray} from '@enact/core/util';
 import intersection from 'ramda/src/intersection';
 import last from 'ramda/src/last';
 
-import {matchSelector} from './utils';
+import {contains, matchSelector, getContainerRect, getRect} from './utils';
 
 const containerAttribute = 'data-spotlight-id';
 const containerConfigs   = new Map();
@@ -60,11 +60,16 @@ let GlobalConfig = {
 	},
 	leaveFor: null,         // {left: <extSelector>, right: <extSelector>, up: <extSelector>, down: <extSelector>}
 	navigableFilter: null,
+	obliqueMultiplier: 5,
+	onEnterContainer: null,     // @private - notify the container when entering via 5-way
+	onLeaveContainer: null,      // @private - notify the container when leaving via 5-way
+	onLeaveContainerFail: null,  // @private - notify the container when failing to leave via 5-way
 	overflow: false,
 	rememberSource: false,
 	restrict: 'self-first', // 'self-first', 'self-only', 'none'
 	selector: '',           // can be a valid <extSelector> except "@" syntax.
 	selectorDisabled: false,
+	straightMultiplier: 1,
 	straightOnly: false,
 	straightOverlapThreshold: 0.5,
 	tabIndexIgnoreList: 'a, input, select, textarea, button, iframe, [contentEditable=true]'
@@ -691,22 +696,35 @@ function getContainerNavigableElements (containerId) {
 	}
 
 	const config = getContainerConfig(containerId);
+	const {enterTo, overflow} = config;
 
-	const enterLast = config.enterTo === 'last-focused';
-	const enterDefault = config.enterTo === 'default-element';
+	const enterLast = enterTo === 'last-focused';
 	let next;
 
 	// if the container has a preferred entry point, try to find it first
 	if (enterLast) {
 		next = getContainerLastFocusedElement(containerId);
-	} else if (enterDefault) {
+	}
+
+	// try default element if last focused can't be focused
+	if (!next) {
 		next = getContainerDefaultElement(containerId);
 	}
 
-	// if there isn't a preferred entry or it wasn't found, try to find something to spot
 	if (!next) {
-		next = (!enterDefault && getContainerDefaultElement(containerId)) ||
-				getSpottableDescendants(containerId);
+		const spottables = getSpottableDescendants(containerId);
+
+		// if there isn't a preferred entry on an overflow container, try to find the first element
+		// in view
+		if (overflow) {
+			const containerRect = getContainerRect(containerId);
+			next = spottables.find(element => contains(containerRect, getRect(element)));
+		}
+
+		// otherwise, return all spottables within the container
+		if (!next) {
+			next = spottables;
+		}
 	}
 
 	return next ? coerceArray(next) : [];
@@ -727,7 +745,6 @@ function getContainerFocusTarget (containerId) {
 	restoreLastFocusedElement(containerId);
 
 	let next = getContainerNavigableElements(containerId);
-
 	// If multiple candidates returned, we need to find the first viable target since some may
 	// be empty containers which should be skipped.
 	return next.reduce((result, element) => {
@@ -946,6 +963,82 @@ function isWithinOverflowContainer (target, containerIds = getContainersForNode(
 		.some(config => config && config.overflow);
 }
 
+/**
+ * Notifies any affected containers that focus has left one of their children for another container
+ *
+ * @param {String} direction up/down/left/right
+ * @param {Node} current currently focused element
+ * @param {String[]} currentContainerIds Containers for current
+ * @param {Node} next To be focused element
+ * @param {String[]} nextContainerIds Containers for next
+ * @private
+ */
+function notifyLeaveContainer (direction, current, currentContainerIds, next, nextContainerIds) {
+	currentContainerIds.forEach(containerId => {
+		if (!nextContainerIds.includes(containerId)) {
+			const config = getContainerConfig(containerId);
+
+			if (config && config.onLeaveContainer) {
+				config.onLeaveContainer({
+					type: 'onLeaveContainer',
+					direction,
+					target: current,
+					relatedTarget: next
+				});
+			}
+		}
+	});
+}
+
+/**
+ * Notifies any containers that focus attempted to move but failed to find a target
+ *
+ * @param {String} direction up/down/left/right
+ * @param {Node} current currently focused element
+ * @param {String[]} currentContainerIds Containers for current
+ * @private
+ */
+function notifyLeaveContainerFail (direction, current, currentContainerIds) {
+	currentContainerIds.forEach(containerId => {
+		const config = getContainerConfig(containerId);
+
+		if (config && config.onLeaveContainerFail) {
+			config.onLeaveContainerFail({
+				type: 'onLeaveContainerFail',
+				direction,
+				target: current
+			});
+		}
+	});
+}
+
+/**
+ * Notifies any affected containers that one of their children has received focus.
+ *
+ * @param {String} direction up/down/left/right
+ * @param {Node} previous Previously focused element
+ * @param {String[]} previousContainerIds Containers for previous
+ * @param {Node} current Currently focused element
+ * @param {String[]} currentContainerIds Containers for current
+ * @private
+ */
+function notifyEnterContainer (direction, previous, previousContainerIds, current, currentContainerIds) {
+	currentContainerIds.forEach(containerId => {
+		if (!previousContainerIds.includes(containerId)) {
+			const config = getContainerConfig(containerId);
+
+			if (config && config.onEnterContainer) {
+				config.onEnterContainer({
+					type: 'onEnterContainer',
+					direction,
+					target: current,
+					relatedTarget: previous
+				});
+			}
+		}
+	});
+}
+
 export {
 	// Remove
 	getAllContainerIds,
@@ -975,6 +1068,9 @@ export {
 	isNavigable,
 	isWithinOverflowContainer,
 	mayActivateContainer,
+	notifyLeaveContainer,
+	notifyLeaveContainerFail,
+	notifyEnterContainer,
 	removeAllContainers,
 	removeContainer,
 	rootContainerId,