+++ packages/moonstone/CHANGELOG.md
@@ -4,12 +4,21 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ## [unreleased]
 
+### Fixed
+
+- `moonstone/Scroller`, `moonstone/VirtualList.VirtualGridList`, and `moonstone/VirtualList.VirtualList` to notify user when scrolling is not possible via voice command
+- `moonstone/TimePicker` to not read out meridiem label when changing the value
+
+## [2.2.0] - 2018-10-02
+
 ### Added
 
 - `moonstone/GridListImageItem` voice control feature support
 
 ### Fixed
 
+- `moonstone/DayPicker` to prevent closing when selecting days via voice control
+- `moonstone/VideoPlayer` to unfocus media controls when hidden
 - `moonstone/Scroller` to set correct scroll position when an expandable child is closed
 - `moonstone/Scroller` to prevent focusing children while scrolling
 
@@ -23,6 +32,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/Scroller` to correctly set scroll position when nested item is focused
 - `moonstone/Scroller` to not adjust `scrollTop` when nested item is focused
 - `moonstone/VideoPlayer` to show correct playback rate feedback on play or pause
+- `moonstone/VirtualList.VirtualGridList` and `moonstone/VirtualList.VirtualList` to handle 5way navigation properly when `focusableScrollbar` is true
 
 ## [2.1.3] - 2018-09-10